Implement dynamic rendering for MarkdownViewer
