+++ examples/Demo/FluentUI.Demo.Client/Layout/DemoMainLayout.razor
@@ -8,54 +8,7 @@
 	@*       Header        *@
 	@* ------------------- *@
 	<FluentLayoutItem Area="@LayoutArea.Header" Height="44px">
-		<FluentStack VerticalAlignment="@VerticalAlignment.Center" HorizontalGap="4px" Height="100%">
-			@* Hamburger on Mobile *@
-			<FluentLayoutHamburger Style="margin: 2px 12px;" PanelHeader="Fluent UI Blazor" Display="HamburgerDisplay.DesktopMobile" />
-
-			@* Microsoft Logo *@
-			<MicrosoftLogo />
-
-			<FluentSpacer />
-
-			@* GitHub Logo *@
-			<FluentButton IconStart="@(new GitHubIcon().WithColor("var(--colorNeutralForegroundOnBrand)"))"
-						  OnClick="@(() => OpenUrlInNewTabAsync("https://github.com/microsoft/fluentui-blazor/tree/dev-v5"))"
-						  Appearance="ButtonAppearance.Transparent"
-						  Title="See the Fluent UI Blazor GitHub repository" />
-
-			@* DarkTheme *@
-			<FluentButton IconStart="@(new Icons.Regular.Size20.DarkTheme().WithColor("var(--colorNeutralForegroundOnBrand)"))"
-						  OnClick="@SwitchThemeAsync"
-						  Appearance="ButtonAppearance.Transparent"
-						  Title="Switch to Light/Dark theme" />
-
-			@* Direction *@
-			<FluentButton IconOnly="true"
-						  Title="Switch to LTR/RTL direction"
-						  OnClick="@SwitchDirAsync"
-						  Appearance="ButtonAppearance.Transparent">
-				<FluentIcon Class="hidden-if-ltr"
-							Value="@(new Icons.Regular.Size20.TextDirectionHorizontalRtl().WithColor(SystemColors.Neutral.ForegroundOnBrand))" />
-				<FluentIcon Class="hidden-if-rtl"
-							Value="@(new Icons.Regular.Size20.TextDirectionHorizontalLtr().WithColor(SystemColors.Neutral.ForegroundOnBrand))" />
-			</FluentButton>
-
-			@* Reboot *@
-			<FluentButton IconStart="@(new Icons.Regular.Size20.DrawText().WithColor("var(--colorNeutralForegroundOnBrand)"))"
-						  OnClick="@ReloadReboot"
-						  Appearance="ButtonAppearance.Subtle"
-						  BackgroundColor="@(_useReboot ? "var(--colorBrandBackground)" : "var(--colorBrandBackgroundHover)")"
-						  Style="display: none;"
-						  Title="Use the Reboot.css styles" />
-
-			@* Open Console *@
-			<FluentButton IconStart="@(new Icons.Regular.Size20.WindowConsole().WithColor("var(--colorNeutralForegroundOnBrand)"))"
-						  OnClick="@(() => _consoleLogOpened = !_consoleLogOpened)"
-						  Appearance="ButtonAppearance.Subtle"
-						  BackgroundColor="@(_consoleLogOpened ? "var(--colorBrandBackground)" : "var(--colorBrandBackgroundHover)")"
-						  Title="Open or hide the console" />
-
-		</FluentStack>
+		<DemoHeader @rendermode="@DocViewer.DemoRenderMode" />
 	</FluentLayoutItem>
 
 	@* ------------------- *@
@@ -64,7 +17,7 @@
 
 	<FluentLayoutItem Area="@LayoutArea.Navigation" Width="262px" Sticky="true" Height="var(--layout-body-height)"
 					  Style="overflow-y: auto; padding: 0px; border-right: var(--strokeWidthThin) solid var(--colorNeutralStroke1);">
-		<DemoNav />
+		<DemoNav @rendermode="@DocViewer.DemoRenderMode" />
 	</FluentLayoutItem>
 
 	@* ------------------- *@
@@ -76,7 +29,8 @@
 		<FluentLayoutItem Area="@LayoutArea.Content" Class="demo-home">
 			<div class="demo-home">
 				<div>@Body</div>
-				<div><DemoAsidePanel Page="@Navigation.ToBaseRelativePath(Navigation.Uri)" /></div>
+				<div><DemoAsidePanel Page="@Navigation.ToBaseRelativePath(Navigation.Uri)" 
+										   @rendermode="@DocViewer.DemoRenderMode" /></div>
 			</div>
 		</FluentLayoutItem>
 	}
@@ -89,7 +43,8 @@
 		</FluentLayoutItem>
 
 		<FluentLayoutItem Area="@LayoutArea.Aside" Width="240px" Sticky="true" hidden-when="sm">
-			<DemoAsidePanel Page="@Navigation.ToBaseRelativePath(Navigation.Uri)" />
+			<DemoAsidePanel Page="@Navigation.ToBaseRelativePath(Navigation.Uri)"
+							@rendermode="@DocViewer.DemoRenderMode" />
 		</FluentLayoutItem>
 	}
